Refactored main module to split across multiple modules for cleanliness.
